refactor: simplify tests
